+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.3] - 2018-12-04
 ### Added
+- `SQLiteDatabase#delete()` method
 
 ## [0.2.2] - 2018-12-04
 ### Added